requirements.txt
streamlit
pandas
beautifulsoup4
python-docx